requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
djangorestframework==3.14.0
idna==3.6
joblib==1.4.2
mysqlclient==2.1.1
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.5.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
